feat: 검색 결과 클릭 시 해당 상세화면으로 이동하도록 구현
#S15P11D207-225

+++ frontend/modera/feature/home/src/main/kotlin/com/ssafy/modera/feature/home/component/HomeBottomSection.kt
@@ -23,7 +23,7 @@ internal fun HomeBottomSection(
     onRecentSearchClick: (String) -> Unit,
     onRecentSearchDelete: (String) -> Unit,
     onSearchDeactivate: () -> Unit,
-    onSearchResultClick: (Int) -> Unit,
+    onSearchResultClick: (Long) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     val focusManager = LocalFocusManager.current
@@ -70,7 +70,7 @@ private fun HomeSearchContent(
     uiState: HomeUiState.Success,
     onRecentSearchClick: (String) -> Unit,
     onRecentSearchDelete: (String) -> Unit,
-    onSearchResultClick: (Int) -> Unit,
+    onSearchResultClick: (Long) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     when {
